add infrastructure for reading and parsing a config file
For simplicity the config file syntax comes from ast.literal_eval(),
which evaluates text as Python literals.  Thus, anything that can be
expressed as a single Python literal object can be used as the
contents of the config file.

Future commits will add semantics to the object returned from parsing
the config file, and thus restrictions on the kinds of objects that
can be expressed in a config file.