add infrastructure for processing config data
This assumes that the config data is a dict object mapping a name
(string) to an appropriate Python literal.  The config processing code
iterates over each name in the dict's set of keys and calls the
registered handler for that name.  If there is no registered handler,
a warning is printed and processing continues.  Handlers are
registered via a config_handler() decorator.

There are no handlers yet -- this is just infrastructure.