create packet capture threads
These threads create a new pcap handle and start capturing.  Each time
a packet is received, the packet's description is extracted and the
appropriate Dumpfile is fetched.  The packet's contents are then
written to the Dumpfile.

Multiple threads are necessary because the user might want to capture
on multiple interfaces.  The pcap API doesn't support opening multiple
interfaces at the same time, hence threads.